+++ src/components/layout/Header.tsx
@@ -6,8 +6,6 @@ import { useEffect, useState } from 'react';
 import HamMenuIcon from '@/assets/icons/HamMenu.svg?react';
 import HamMenuMIcon from '@/assets/icons/HamMenuM.svg?react';
 import HamMenuTIcon from '@/assets/icons/HamMenuT.svg?react';
-import { postSignout } from '@/apis/auth/postSignout';
-import { useHandleUnauthorized } from '@/utils/handleUnauthorized';
 import { useRecoilValue } from 'recoil';
 import { headerState } from '@/recoil/header';
 
@@ -15,7 +13,6 @@ const Header = () => {
   const { isMobileOrTablet, isTablet, isMobile, isDesktop } = useNSMediaQuery();
   const [isViewHamItem, setIsViewHamItem] = useState(false);
   const isAccessToken = !!localStorage.getItem('accessToken');
-  const handleUnauthorized = useHandleUnauthorized();
   const navigate = useNavigate();
   const header = useRecoilValue(headerState);
 
@@ -25,25 +22,12 @@ const Header = () => {
     }
   }, [isDesktop]);
 
-  const handleSignOut = async () => {
-    if (isAccessToken) {
-      const response = await postSignout(handleUnauthorized);
-      if (response.check) {
-        console.log('로그아웃');
-        localStorage.removeItem('accessToken');
-        localStorage.removeItem('refreshToken');
-      }
-    }
-  };
-
   return (
     <>
       {isMobileOrTablet ? (
         <>
           <S.Container>
-            <S.NavMain to="/" onClick={handleSignOut}>
-              N력 키우기
-            </S.NavMain>
+            <S.NavMain to="/">N력 키우기</S.NavMain>
             <HamMenuIcon onClick={() => setIsViewHamItem(!isViewHamItem)} />
           </S.Container>
           {isViewHamItem && (
@@ -92,9 +76,7 @@ const Header = () => {
         </>
       ) : (
         <S.Container>
-          <S.NavMain to="/" onClick={handleSignOut}>
-            N력 키우기
-          </S.NavMain>
+          <S.NavMain to="/">N력 키우기</S.NavMain>
           <S.Nav>
             <S.NavLinksMain
               $isActive={header === 'service'}
